+++ libraries/ui/CMakeLists.txt
@@ -1,3 +1,3 @@
 set(TARGET_NAME ui)
-setup_hifi_library(OpenGL Network Qml Quick Script WebChannel WebEngine WebSockets XmlPatterns)
+setup_hifi_library(OpenGL Network Qml Quick Script WebChannel WebSockets XmlPatterns)
 link_hifi_libraries(shared networking gl)
